T: use a separate package instead of Category for slow tests
